source-pipeline/requirements.txt
cassandra-driver==3.13
requests